Time: 0 ms (100.00%), Space: 5.9 MB (93.20%) - LeetHub

+++ 0371-sum-of-two-integers/0371-sum-of-two-integers.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int getSum(int a, int b) {
+    short getSum(short a, short b) {
         return a+b;
     }
 };
